+++ apps/docs/src/content/04-components/form-controls/password-creation-field/examples/default.tsx
@@ -1,7 +1,5 @@
-import {
-  Label,
-  PasswordCreationField,
-} from "@mittwald/flow-react-components";
+import { Label } from "@mittwald/flow-react-components";
+import { PasswordCreationField } from "@mittwald/flow-react-components/password-tools";
 import { useState } from "react";
 
 export default () => {